Enumerate saml roles from assertion rather than okta api
The roles associated with a given Okta AWS application are not quite
accurate as they should specify the idp they expect to associate with.

This data can be extracted from the SAML assertion rather than just
trusting the list of roles returned by the OKTA API enumeration.

The issue is discussed further here -
https://github.com/Nike-Inc/gimme-aws-creds/issues/18

This pull request makes the following changes -

- Silly edit to gitignore to ignore Vim temp files

In bin/gimme-aws-creds
- base65/xml libraries for parsing saml assertion and extracting roles
- New function _enumerate_saml_roles which will extract all idp
  provider arn / role arn pairs from a saml assertion and return list
  of named tuples
- modified choose_role to enumerate and select from list of named
  tuples not from okta response
- modified behavior of credential gathering so that all roles can be
  enumerated.  this allows aws users with multiple roles across
  multiple accounts to have all their 1h credentials collected.

I believe all the original functionality is still intact, but
would definitely like another set of eyes to review and test!