Updated the handling of MFA failures to prevent exceptions and print useful error messages instead.
Added ability to specify a MFA code (for TOTP and SMS) on the command line with a new --mfa-code parameter.  If the parameter isn't provided, the existing prompt is still used to query the user.

Cosmetic update to the main.py comment to include the current command line parameters.

This allows the gimme-aws-creds CLI tool to be used by other scripts that won't necessarily be able to prompt the user for MFA code input.  With other CLI tools, like oathtool, MFA can be handled in an automated manner for trusted devices that can generated there MFA codes automatically on demand.