Fix issue where the prompt for saving the password in the keyring was not being printed to stderr.
There was an earlier patch that converted all screen output to stderr so that credentials printed on stdout could be redirected to a file, however when using the script in this way, the save password prompt was obscured from the user making it appear that the script had hung, as well as polluting the redirected output with the prompt text.  This change prints input prompts to sys.stderr to fix the issue.  There may be other input prompts that have this issue, as I did not audit the remaining code outside of the okta.py script.