Aliases for webauthn authenticators.
Building on the addition to registered authenticators, this adds support
for adding, and using, aliases on authenticators.

There are two points where we register an alias, the first is on device
registration, and the second is on successful use.

On the successful use case, we check to see if we already have an alias
on the authenticator in question, and if we do not, we prompt for one.

This makes it possible to track webauthn authenticators which were not
registered via gimme_aws_creds.