Add an optional use_keyring config option.
If absent, or y, then we will use a system keyring if available.

If it is n, then we will not use a system keyring even if it is
available.

This should prevent continual prompts for storing it in the system
keyring for people who either do not wish to store it there, or for
people whose corporate policies do not allow them to store it there.