Support for passing a list of credential IDs.
If we get a list of IDs, then we pass the whole list as the accept list,
and if any of them are supported by a given device we can use it.

This is required for being able to just ask the user if they want to use
webauthn, instead of asking them to pick which webauthn device from a
potentially unlabeled list that they want to use.