Rework except logic a little bit.
This is almost entirely to quiet the linter, however it also simplifies
the code a little bit, and adds a comment to explain the code.

Specifically, we should either check each and every layer of the object
tree to see if the components exist, or we should rely on the
AttributeError exception, not a combination of both.

And we should only catch AttributeError, so we don't mask code problems
in self._print_correct_answer.