Consider N Webauthn factors as a single option.
Currently, if a user has 5 Webauthn keys configured in Okta (or, more
likely, 2), the user gets 5 separate options in the factor list.

This is...  Not horribly user friendly, even if every single one does
have a name configured for them.

It turns out that Okta lets us ask for a nonce valid for every single
one of them.

If we take that path, we can take that nonce, and the whole list of
possible challenge IDs, and pass it down to the webauthn code.

At that point, we can let it pick whichever device is actually plugged
in to this computer at this time.

This means both that we can just give a single option in the factor list
for Webauthn, and that setting the preferred MFA type to webauthn can
Just Work even when a user has multiple webauthn tokens configured.

This makes for a much better UI flow for users.

The only big outstanding question I have is if the code around having
names / labels for webauthn tokens makes any sense after this.