fido2 1.0.0 support.
This makes the changes required to support the 1.0.0 and above releases
of fido2.

The biggest two changes are in having to implement
fido2.client.UserInteraction instead of just passing in the PIN and an
on_keepalive function callback, and in having to upgrade to a version of
ctap-keyring-device which also supports fido2 1.0.0.

For ctap-keyring-device, we have the complication that while 2.0.0
supports fido2 1.0.0, it seems that the 2.0.0 release hasn't actually
been, well, released yet.

As such, we're quite improperly importing directly from git right now.

We should fix that.