Attempt to keep and use the same Okta session.
This is a small code change, and a huge functionality change.

And it is only supported in 'internal' and 'appurl'
gimmie_creds_server modes.

But with a successful authorization via Okta, we will now store the
username and Okta session cookie in the config file as session_username
and session_token.

When we go to authorize, we load both from the config, and if the
username is either unset, or it matches what is in session_username, we
will attempt to renew the session cookie.

If the Okta session is still valid, that will extend it by an Okta
configured duration, and we then skip the entire authorization flow and
just use that Okta session.

If the session is not valid, we will get a 404 back, and we continue
forward with the standard flow.

There is a significant catch with this, if you want to change Okta
usernames, you must specify the username on the command line.