Rework keyring devices a little.
Two things, first, we now always attempt to get the list of keyring
devices, and if we are successful, we concat the two lists of devices.

Second, we catch a known error in CtapKeyringDevice, and if we encounter
it we print a message pointing at the PR which should fix the error.