Rework _run_in_thread error handling.
self.locate_device() never raises an error, so we need to check for
_clients being empty ourselves, and if it is still empty after the
second try, we raise the error ourselves.

There is now a self._exception, which is a list of exceptions.
Functions run in threads are expected to append any errors to this list.

At the end, we check the list of errors.

We default to FIDODeviceTimeoutError('Operation timed out')

We then loop through the exception list, if we have instances of only
FIDODeviceTimeoutError, we keep the default exception, if we find an
instance of NoEligibleFIDODeviceFoundError, we switch to an error of
NoEligibleFIDODeviceFoundError('No eligible authentication devices
found.'), and if we have any other exception on the list, we pick that
one instead over all of the others, wrapping it in a FIDODeviceError.

We then throw whatever error we ended up with.

These errors are now caught by the calling okta code.

This should vastly improve the user experience around errors.