Rewrite make_credential.
With current versions of python-fido2, this needed a fair bit of
attention, and while I was at it, I started actually paying attention to
a lot more of what Okta is sending us in regards to parameters for the
credentials.

There is also a workaround for a python-fido2 bug here, with a comment
that has a link to the bug report.